StudentProfilePage change to organism and add PasswordChange molecule.
