+++ src/components/molecules/PasswordChange/PasswordChange.tsx
@@ -1,21 +1,11 @@
-import {PrimaryButton} from "../../atoms/PrimaryButton/PrimaryButton";
 import {PrimaryInput} from "../../atoms/PrimaryInput/PrimaryInput";
-import { StudentProfileStyle } from "./StudentProfile.style";
+import {PrimaryButton} from "../../atoms/PrimaryButton/PrimaryButton";
 import {useState} from "react";
 
-export const StudentProfile = () => {
+export const PasswordChange = () => {
 
     const [passwordInput, setPasswordInput] = useState('');
     const [repeatPasswordInput, setRepeatPasswordInput] = useState('');
-
-    const gotAJobHandler = () => {
-
-        if(window.confirm("Jesteś pewien? Ta zmiana spowoduje usunięcie twojego konta z listy kursantów")){
-            console.log('Usuwam!');
-        }else{
-            console.log('Cofam!');
-        }
-    }
 
     const changePasswordHandler = () => {
         if(!passwordInput) {
@@ -34,10 +24,7 @@ export const StudentProfile = () => {
     }
 
     return(
-        <StudentProfileStyle>
-            <p>Zatrudnienie</p>
-            <PrimaryButton text={"Zostałem zatrudniony!"} onClick={gotAJobHandler}/>
-            <hr/>
+        <>
             <br/>
             <p>Zmień hasło</p>
             <PrimaryInput type={"password"} placeholder={'Wpisz hasło'} onChange={(e)=>setPasswordInput(e.target.value)}/>
@@ -45,6 +32,6 @@ export const StudentProfile = () => {
             <PrimaryInput type={"password"} placeholder={'Powtórz hasło'} onChange={(e)=>setRepeatPasswordInput(e.target.value)}/>
             <br/>
             <PrimaryButton text={"Zatwierdź"} onClick={changePasswordHandler}/>
-        </StudentProfileStyle>
-        )
+        </>
+    )
 }
