+++ src/components/organisms/StudentProfile/StudentProfile.style.ts
